new fitness function with Wilson score

+++ src/minijava/comparator/FitnessComparatorByCombined.java
@@ -13,7 +13,8 @@ public class FitnessComparatorByCombined implements Comparator<Fitness> {
 			if(compare == 0) {
 				compare = Integer.compare(fitness1.size, fitness2.size);
 				if(compare == 0) {
-					compare = Long.compare(fitness2.meanCorrectConfidenceIntervalScaled, fitness1.meanCorrectConfidenceIntervalScaled);		// flip order to obtain largest correct first
+					// flip order to obtain largest correct first
+					compare = Double.compare(fitness2.meanCorrectConfidenceIntervalScaled, fitness1.meanCorrectConfidenceIntervalScaled);
 					if(compare == 0) {
 						compare = fitness1.meanErrorConfidenceIntervalScaled.compareTo(fitness2.meanErrorConfidenceIntervalScaled);
 					}
