+++ src/minijava/comparator/FitnessComparatorByMeanErrorConfidenceInterval.java
@@ -15,7 +15,8 @@ public class FitnessComparatorByMeanErrorConfidenceInterval implements Comparato
 				if(compare == 0) {
 					compare = fitness1.meanErrorScaled.compareTo(fitness2.meanErrorScaled);
 					if(compare == 0) {
-						compare = Long.compare(fitness2.meanCorrectScaled, fitness1.meanCorrectScaled);		// flip order to obtain largest correct first
+						// flip order to obtain largest correct first
+						compare = Double.compare(fitness2.meanCorrectScaled, fitness1.meanCorrectScaled);
 					}
 				}
 			}
